Removed all traces of proprietary files from commit history
All commits prior to this one have been rewritten to remove all
proprietary files.  (If you cloned this repository prior to this point,
you will have to rebase your branches.)

This was done with the purpose of removing any legal issues from this
repository.  It can now be freely distributed without any concern.

If you want to recreate a previous commit for some reason, simply unpack
the eawpatches archive over the commit tree -- the *.pat files should be
directly in the root directory.  (This of this like svn-buildpackage's
mergeWithUpstream mode.)